Control.HException

Description

Heterogeneously typed exceptions.

Documentation

data HException es

A heterogeneously typed Exception, which holds an underlying exception that is statically constrained to be one of the type-level list es.

type (:^:) e es = Tagged e e : es infixr 7

Type alias for ease of constructing exception type-level lists.

type Only e = e :^: `[]`

A singleton exception type-level list.

type HException1 e = HException (Only e)

The type of HExceptions containing only one type of exception.

type Both es es' = HAppendListR es es'

The type of heterogeneous exceptions consisting of both es and es'.

type TypeIndexed xs = HTypeIndexed xs

The type-level list xs is constrained to contain only distinct types, and so can be indexed via type alone.

type Member x xs = MkVariant x x xs

The type x is contained in the type-level list xs.

hException :: (Member e es, TypeIndexed es) => e -> HException es

Construct a HException from a standard exception value.

getMay :: Member e es => HException es -> Maybe e

Extract an exception of a specific type from a HException. Returns Nothing if the exception encased is not of the required type.

get :: HException1 e -> e

Specialization of getMay for cases where there is only a single type of exception.

type Subset xs xs' = ExtendsVariant xs xs'

Constrain that the type-level list xs' contains all of the elements of the type-level list xs.

generalize :: (Subset es es', TypeIndexed es') => HException es -> HException es'

Changes the type of a HException so it is more general than the original, without changing the underlying exception.