Control.HException

Description

Heterogeneously typed exceptions.

Documentation

data HException es

A heterogeneously typed Exception, which holds an underlying exception that is statically constrained to be one of the type-level list es.

type (:^:) e es = Tagged e e : es infixr 7

Type alias for ease of constructing exception type-level lists.

type Only e = e :^: `[]`

A singleton exception type-level list.

type HException1 e = HException (Only e)

The type of HExceptions containing only one type of exception.

type family Concat xs :: [*]

Concatenate a type-level list of lists (useful for combining two or more heterogeneous exception types).

Equations

Concat `[]` = `[]`

Concat (x : xs) = HConcatR (HList x : (HList (Concat xs) : `[]`))

type TypeIndexed xs = HTypeIndexed xs

The type-level list xs is constrained to contain only distinct types, and so can be indexed via type alone.

type Member x xs = MkVariant x x xs

The type x is contained in the type-level list xs.

hException :: (Member e es, TypeIndexed es) => e -> HException es

Construct a HException from a standard exception value.

getMay :: Member e es => HException es -> Maybe e

Extract an exception of a specific type from a HException. Returns Nothing if the exception encased is not of the required type.

get :: HException1 e -> e

Specialization of getMay for cases where there is only a single type of exception.

type Slice xs xs' xs'' = (SplitVariant xs xs' xs'', DeleteAll xs' xs xs'')

Split xs into two groups xs' and xs''.

slice :: (Slice es el er, TypeIndexed el, TypeIndexed er) => HException es -> Either (HException el) (HException er)

Extract either the exceptions in the left type-level list el if the type of the underlying exception is in el, or otherwise return an exception in the right type-level list.

type Subset xs xs' = ExtendsVariant xs xs'

Constrain that the type-level list xs' contains all of the elements of the type-level list xs.

generalize :: (Subset es es', TypeIndexed es') => HException es -> HException es'

Changes the type of a HException so it is more general than the original, without changing the underlying exception.

class TransformSubset i sub sub' o | i sub -> sub' where

Methods

transformSubset :: Applicative m => (HException sub -> m (HException o)) -> HException i -> m (HException o)

Apply a function to transform a subset (sub) of the possible exception types contained in i, if any other exception is encountered then it is carried through to the output without transformation.

mapSubset :: TransformSubset i sub sub' o => (HException sub -> HException o) -> HException i -> HException o

Map a subset of the exception types in the input using the provided function, any other exception types will be carried through to the output.