Control.Monad.Trans.HExcept

Description

Functions for handling HExceptions using ExceptT.

Documentation

type HExcept es = Except (HException es)

The heterogeneous exception monad. Computations are either one of a known set of exceptions es or a value.

type HExcept1 e = HExcept (Only e)

A specialization of HExcept for computations which can return only one type of exception.

type HExceptT es = ExceptT (HException es)

A monad transformer that adds heterogeneous exceptions into other monads.

type HExceptT1 e = HExceptT (Only e)

A specialization of HExceptT for computations which can return only one type of exception.

hExceptT :: Applicative m => HExcept es a -> HExceptT es m a

Lift a HExcept into its transformer version.

extend :: (Functor m, Subset es es', TypeIndexed es') => HExceptT es m a -> HExceptT es' m a

Extend the given result so that it may be used in a context which can return a superset of the errors that may arise from the original result. This is useful for calling multiple functions which return different errors types from within a single function. Note: this can also be used to re-order the error types.

hThrowE :: (Member e es, Monad m, TypeIndexed es) => e -> HExceptT es m a

Signal an exception value e.

type Handler es es' a = HandlerT es es' Identity a

The type of functions used to handle the exceptions es, which may return the exceptions es'.

type HandlerT es es' m a = HException es -> HExceptT es' m a

The monad transformer version of Handler.

handler1 :: (e -> HExceptT es m a) -> HandlerT (Only e) es m a

Construct a handler from a function which operates on the underlying exception type.

handlerT :: Applicative m => Handler es es' a -> HandlerT es es' m a

Lift a (pure) handler into m.

orElse :: (Slice es'' es es', TypeIndexed es, TypeIndexed es') => HandlerT es os m a -> HandlerT es' os m a -> HandlerT es'' os m a infixr 2

Chain HandlerTs together.

orDefault :: ProjectVariant es' es => HandlerT es os m a -> HExceptT os m a -> HandlerT es' os m a

Try to use the provided HandlerT or, if that handler does not apply, return the default.

type Value = HExcept `[]`

A specialization of HExcept for computations in which all possible exception types have been dealt with. This means the domain of possible exceptions is empty.

type ValueT = HExceptT `[]`

The monad transformer version of Value.

value :: Value a -> a

Extract the value from a computation result after all exceptions have been handled. This is type-safe because the type of exceptions is constrained to be empty.

valueT :: Monad m => ValueT m a -> m a

Monad transformer version of value.